Move WS handle to home/index
