+++ pages/home/index.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux'
+import { useSelector, useDispatch } from 'react-redux'
 
 import { RootReducer } from '@/store';
 
@@ -8,8 +8,10 @@ import Login from '@/pages/home/login';
 
 import useWebSocket from 'react-use-websocket';
 import { URI } from "@/store/server_urls";
+import { setRoom, setPlayer, leaveMatch, setMatch } from "@/store/reducers/matchReducer"
 
 export default function HomeScreen() {
+    const dispatch = useDispatch()
     const userData = useSelector((state: RootReducer) => state.authReducer.user_data)
     const [ws_url, setWsUrl] = useState("")
 
@@ -38,6 +40,29 @@ export default function HomeScreen() {
         }
     }, [userData])
 
+    useEffect(() => {
+        if (WS.lastJsonMessage) {
+            const data = WS.lastJsonMessage as APIResponseProps
+            console.log('<<<<< RESP: ', data)
+            if (data.data_type === "room_update") {
+                console.log('room_update')
+                dispatch(setRoom(data.room_data!))
+            }
+            else if (data.data_type === "player_update") {
+                console.log('player_update')
+                dispatch(setPlayer(data.player_data!))
+            }
+            else if (data.data_type === "disconnected") {
+                console.log('disconnected')
+                dispatch(leaveMatch())
+            }
+            else if (data.data_type === "match_update") {
+                console.log('match_update')
+                dispatch(setRoom(undefined))
+                dispatch(setMatch(data.match_data!))
+            }
+        }
+    }, [WS.lastJsonMessage])
 
     if (userData) {
         return <Home />;
